test(#686): address review — split helpers, dedupe login, rename fixture
