+++ apps/web-e2e/specs/avatar-satellite.spec.ts
@@ -1,5 +1,5 @@
+import { runLogin } from '../src/support/run-login';
 import { expect, test } from '../src/support/test';
-import { waitForHoneypotWindow } from '../src/wait-for-honeypot-window';
 
 /**
  * `/avatar` mounts its own satellite canvas alongside the persistent world canvas: two `<canvas>`
@@ -21,20 +21,9 @@ test(
       }
     });
 
-    await page.setExtraHTTPHeaders({ 'x-forwarded-for': '127.0.0.1' });
+    await runLogin(page, { email: 'e2e-avatar-satellite@vers.test', password: 'password123' });
+
     await page.goto('/avatar');
-
-    await expect(page).toHaveURL(/\/login/);
-
-    // hydration gate: the login form's submit handler attaches only once React commits; an
-    // earlier click falls back to the browser's native GET submit and never leaves /login
-    await page.locator('html[data-hydrated]').waitFor();
-    await page.getByLabel('Email').fill('e2e-avatar-satellite@vers.test');
-    await page.getByLabel('Password').fill('password123');
-
-    await waitForHoneypotWindow(page);
-
-    await page.getByRole('button', { exact: true, name: 'Login' }).click();
 
     await expect(page).toHaveURL(/\/avatar$/);
