+++ apps/web-e2e/specs/signup-journey.spec.ts
@@ -1,16 +1,14 @@
-import {
-  buildAvatarName,
-  runLogin,
-  runSettingsLogout,
-  runSignUpIntoGame,
-} from '../src/support/journey';
+import { buildAvatarName } from '../src/support/build-avatar-name';
+import { runLogin } from '../src/support/run-login';
+import { runSettingsLogout } from '../src/support/run-settings-logout';
+import { runSignUpIntoGame } from '../src/support/run-sign-up-into-game';
 import { expect, test } from '../src/support/test';
 import type { JourneyAccount } from '../src/support/types';
 
 test(
   'it signs up, creates an avatar, and can navigate the shell to settings and log out',
   { tag: '@journey' },
-  async ({ getVerificationCode, page }) => {
+  async ({ page, waitForVerificationCode }) => {
     const runID = Date.now();
 
     const account: JourneyAccount = {
@@ -20,7 +18,7 @@ test(
       username: `e2esignup${runID}`,
     };
 
-    await runSignUpIntoGame(page, account, getVerificationCode);
+    await runSignUpIntoGame(page, account, waitForVerificationCode);
     await expect(page.locator('canvas').first()).toBeVisible();
     await runSettingsLogout(page);
     await expect(page.getByRole('link', { name: 'Log in' })).toBeVisible();
@@ -30,7 +28,7 @@ test(
 test(
   'it logs a fresh account out and back in, landing signed in at respite',
   { tag: '@journey' },
-  async ({ getVerificationCode, page }) => {
+  async ({ page, waitForVerificationCode }) => {
     const runID = Date.now();
 
     const account: JourneyAccount = {
@@ -40,8 +38,9 @@ test(
       username: `e2elogin${runID}`,
     };
 
-    await runSignUpIntoGame(page, account, getVerificationCode);
+    await runSignUpIntoGame(page, account, waitForVerificationCode);
     await runSettingsLogout(page);
     await runLogin(page, account);
+    await expect(page).toHaveURL(/\/respite$/);
   },
 );
